+++ lib/page/bookKeeping/keeping_income.dart
@@ -47,7 +47,7 @@ class _KeepIncomeState extends State<KeepIncome> {
   final _timeTextController = TextEditingController();
 
 
-  int _keepText = 0;
+  double _keepText = 0.00;
   String _beiZhuText = '';
   String _timeText = '';
 
@@ -63,7 +63,7 @@ class _KeepIncomeState extends State<KeepIncome> {
   void initState() {
     super.initState();
     setState(() {
-      _keepText = 0;
+      _keepText = 0.00;
       _beiZhuText = _beiZhuTextController.text;
       _timeText = _timeTextController.text;
       curImageNumString = '';
@@ -186,9 +186,9 @@ class _KeepIncomeState extends State<KeepIncome> {
                 print(curImageNumString);
                 print(curCategoryName);
 
-                int _keepText =
-                int.parse(_keepTextController.text) == 0
-                    ? 0
+                double _keepText =
+                double.parse(_keepTextController.text) == 0.00
+                    ? 0.00
                     : int.parse(_keepTextController.text);
                 String _beiZhuText = _beiZhuTextController.text == ''
                     ? '无'
